requirements.txt
# For the web server
gunicorn==22.0.0
Flask==3.0.3

# For audio processing, transcription, and diarization
whisperx
ffmpeg-python==0.2.0

# The following line installs PyTorch with CUDA 11.8 support
-f https://download.pytorch.org/whl/torch_stable.html
torch==2.0.0+cu118
torchvision==0.15.0+cu118
torchaudio==2.0.0+cu118

# For interacting with Google Cloud
google-cloud-storage==2.17.0
google-cloud-secret-manager==2.20.0